King's Men Rework
So this is the other half of reworking the King's Men, splitting them into the Melkorites and the King's Men themselves. The Melkorites are those that are diehard supporters of Melkor, sacrificing humans, etc. However, they have (at least in Umbar) been basically wiped out by its time as a Gondorian province. In comparison, the King's Men have reverted back to their more political form - they are essentially Numenorean supremacists. They dislike the elves and the valar, but they also dislike Sauron and Melkor - anyone they view as attempting to 'control' humans (and the Numenoreans more specifically) . They were able to assimilate the followers of Castamir because of this focus on Numenorean superiority. I do think some sort of 'corsair' religion seperate from the Kings Men might be appropriate for the region.