Mornionath Family Tree- Second Age Complete
Family tree for the Mornionath Family, a family of Lords of Korlea who were fanatically loyal to the King in Armenelos and eventually became Mulkherites, after which they were extinguished near the time of the Downfall. I also added Zigurkhil, so if he is needed for further reference, look no further.