Vineland Guard attire Helm
So first WIP update for https://github.com/MattTheLegoman/RealmsInExile/issues/347 that i am to add all of the vineland guard armour to dorwinion. What's in now is the helm, it's yet tied to a culture (I tried) will need to polish that. but it can be found in the barbershop. some texture refinements might need to be added. it does not have it's properties.dds texture file (since it doesn't exist) hopefully this won't cause an issue (as last time i added a helm everyone became horny)