Bandaid for Elf Holdings
This is a temporary bandaid to make sure Elf always have the Elven government type, it changed their capital province to an elven refuge regardless of its culture since your capital province needs to be a holding your government type can hold, or else it errors out and doesn't change it. I know it's stupid.