ai aggressiveness
This adds an effect that increases ai_boldness, which makes a character more likely to go to war. The existence of this should address #417 , but it may result in them going too aggressive for aragorn's own good. (at 100 at least) I made it a house modifier such that in case the character in question dies, his or her successor will carry on the policy for some time. May also help to fix #940 , or for some other characters.