Update Edhellond Decisions for Numenor-in-Exile
Should now work regardless of your primary title being gondor, reunited kingdom, or numenor in exile.

In theory any top liege with religion:dunedain_religion can enact the decision.  If this is not desired behavior a change will have to be made, but consider that as originally written any direct holder of edhellond with religion:dunedain_religion could get the event, regardless of whether or not they were a part of Gondor, Arnor, RK, or Numenor-in-Exile.