Ulukai claim if it already exists
Somebody reported not getting the Ulukai from this (it was because they didn't have the RC DLC), but in the midst of trying to help them, I realized that it is possible for somebody to send an adventurer and get the Ulukai before the orcs revive Gundabad. In that case, I believe the Gundabad orc ruler should get a claim on the artifact.